datareference/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-dotenv==1.0.0
azure-storage-blob==12.19.0
python-multipart==0.0.6
starlette==0.35.1
pydantic==2.5.3
httpx==0.25.2 